+++ CHANGELOG.md
@@ -3,13 +3,23 @@
 All notable changes to this project will be documented in this file.
 Version number must be set in app.json file.
 
-## [1.0.1] - 2026-03-01
+## [1.0.5] - 2 2026-06-16
 
-- Update: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
+- Fixed auto-login after log out.
+- Removed unnecessary link in Payslip page.
 
-- Fix: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
+## [1.0.4] - 2 2026-06-10
 
-- New feature: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
+- Fixed "My Schedule" bug (mixing schedules from my team members).
+- Enabled PostHog in Training env.
+- Improvement in orion.app flow.
+- Improvement in post image rendering by using skeletons.
+- Defensive fallback to clean tokens and force biometric popup to show always in first login.
+
+## [1.0.3] - 2 2026-05-29
+
+- Visual improvements based on Design Team feedback.
+- New Onboarding page design.
 
 ## [1.0.0] - 2 2026-02-19
 
@@ -18,4 +28,6 @@ First release including:
 - Payslip
 - Schedules
 - Profile
+- Notifications (tasks)
+